Refactor: Update import paths for prisma-json utility
